+++ mammoth/__init__.py
@@ -151,7 +151,7 @@ from mammoth.models.workspaces import (
 )
 from mammoth.view import View, ViewExport
 
-__version__ = "0.7.16"
+__version__ = "0.7.17"
 __all__ = [
     # Client
     "MammothClient",